+++ rclickatell/__init__.py
@@ -1,0 +1,4 @@
+"Clickatell backend for the RapidSMS project."
+
+
+__version__ = '0.1.0'