+++ rclickatell/views.py
@@ -2,18 +2,13 @@ import logging
 import pprint
 
 from django.http import HttpResponse
-from django.conf import settings
 from django.shortcuts import render_to_response
 from django.template import RequestContext
 from django.views.decorators.csrf import csrf_exempt
 
-from rapidsms.models import Contact, Connection
-from rapidsms.messages.outgoing import OutgoingMessage
-from rapidsms.router import router
 from rapidsms.contrib.ajax.utils import request as ajax_request
 
 from rclickatell.forms import MessageForm, StatusCallbackForm
-from rclickatell.models import MessageStatus, Message
 
 
 logging.basicConfig(level=logging.DEBUG)
@@ -47,6 +42,6 @@ def status_callback(request):
         form.save(ip_address=request.get_host())
     else:
         post = pprint.pformat(request.POST)
-        errors = [(k, unicode(v[0])) for k, v in form.errors.items()]
+        errors = [(k, str(v[0])) for k, v in form.errors.items()]
         logging.error('Callback error: %s, %s' % (errors, post))
     return HttpResponse('OK')