+++ RELEASE_NOTES.rst
@@ -1,6 +1,19 @@
 Release Notes
 =============
 
+v0.1.0 - 2015-07-23
+-------------------
+
+* Update setup.py, .gitignore
+* Testing, Flake8, and coverage reporting
+* Travis CI support
+* Python 3 support
+* Django 1.7+ support
+* RapidSMS 0.18+ support
+* This release DROPS support for Python 2.6
+* This release DROPS support for Django < 1.7
+* This release DROPS support for RapidSMS < 0.18
+
 v0.0.3 - 2010-08-25
 -------------------